requirements.txt
pyTelegramBotAPI
google-generativeai
aiohttp
md2tgmd
google-api-python-client
duckduckgo-search>=4.1.1